// src/campaigns/email-campaigns/campaigns/domain/types.ts
export type EmailMetricsCSV = {
  campaignName: string;
  eventType: string;
  from: string;
  to: string;
  timestamp: string;
};

export type CSVHeaderFormat = {
  'Campaign Name': string;
  'Event Type': string;
  From: string;
  To: string;
  Timestamp: string;
};

export type CSVUploaderJob = {
  formattedData: { campaignId: SchemaId; data: EmailMetricsCSV[] }[];
  customer: CustomerDocument;
  email: string;
  bucket: string;
};